add reader and test count fn
