+++ Cargo.toml
@@ -11,6 +11,7 @@ serde = "1"
 serde_derive = "1"
 serde_json = "1"
 failure= "0.1"
+log = "0.4"
 
 [dependencies.chrono]
 features = ["serde"]
